Revert "Moving files to new destinations"
This reverts commit 02e0c42fc1d8b393776e1acf36892c9fe54b3f3d.
