+++ examples/kuka_iiwa_arm/BUILD.bazel
@@ -51,7 +51,7 @@ drake_cc_library(
     deps = [
         ":iiwa_common",
         ":iiwa_lcm",
-        "//manipulation/util:robot_plan_interpolator",
+        "//manipulation/planner:robot_plan_interpolator",
         "//systems/framework:diagram_builder",
         "//systems/primitives:demultiplexer",
     ],
