+++ multibody/inverse_kinematics/BUILD.bazel
@@ -17,23 +17,9 @@ drake_cc_package_library(
     name = "inverse_kinematics",
     visibility = ["//visibility:public"],
     deps = [
-        ":constraint_relaxing_ik",
         ":global_inverse_kinematics",
         ":inverse_kinematics_core",
         ":kinematic_evaluators",
-    ],
-)
-
-drake_cc_library(
-    name = "constraint_relaxing_ik",
-    srcs = ["constraint_relaxing_ik.cc"],
-    hdrs = ["constraint_relaxing_ik.h"],
-    deps = [
-        ":inverse_kinematics_core",
-        "//common/trajectories:piecewise_polynomial",
-        "//multibody/parsing",
-        "//multibody/plant",
-        "//solvers:solve",
     ],
 )
 
@@ -122,19 +108,6 @@ drake_cc_library(
 )
 
 #============ Test ============================
-
-drake_cc_googletest(
-    name = "constraint_relaxing_ik_test",
-    srcs = ["test/constraint_relaxing_ik_test.cc"],
-    data = [
-        "//manipulation/models/iiwa_description:models",
-    ],
-    tags = ["snopt"],
-    deps = [
-        ":constraint_relaxing_ik",
-        "//common:find_resource",
-    ],
-)
 
 drake_cc_library(
     name = "inverse_kinematics_test_utilities",
